Describe changes

+++ mind_space_backend/src/modules/comment/comment.service.js
@@ -1,6 +1,6 @@
 import { Article } from "../../db/models/article.js";
 import { Comment } from "../../db/models/comment.js";
-import { messages } from "../../utils/index.js";
+import { AppError, messages } from "../../utils/index.js";
 
 export const createComment = async (req, res, next) => {
   const { articleId, id } = req.params;
@@ -8,7 +8,7 @@ export const createComment = async (req, res, next) => {
 
   const articleExists = await Article.findById(articleId);
   if (!articleExists) {
-    return next(new Error(messages.article.notFound, { cause: 404 }));
+    return next(new AppError(messages.article.notFound,  404 ));
   }
   const createdComment = await Comment.create({
     article: articleId,
@@ -29,15 +29,24 @@ export const getComments = async (req, res, next) => {
 
   const articleExists = await Article.findById(articleId);
   if (!articleExists) {
-    return next(new Error(messages.article.notFound, { cause: 404 }));
+    return next(new AppError(messages.article.notFound,  404 ));
   }
 
-  const comments = await Comment.find({
+  let comments=[]
+  if(id){
+   comments = await Comment.find({
     article: articleId,
     parentComment: id,
   }).populate([
     { path: "user", select: "userName pfp.secure_url" },
   ]);
+  }else{
+  comments = await Comment.find({
+    article: articleId,
+  }).populate([
+    { path: "user", select: "userName pfp.secure_url" },
+  ]);
+  }
 
   return res.status(200).json({
     success: true,
@@ -50,17 +59,21 @@ export const deleteComment = async (req, res, next) => {
 
   const articleExists = await Article.findById(articleId);
   if (!articleExists) {
-    return next(new Error(messages.article.notFound, { cause: 404 }));
+    return next(new AppError(messages.article.notFound,  404 ));
   }
 
   const comment = await Comment.findById(id);
+  
+  if(!comment){
+    return next (new AppError(messages.comment.notFound,404))
+  }
 
   if (
     ![comment.user.toString(), articleExists.publisher.toString()].includes(
       req.authUser.id,
     )
   ) {
-    return next(new Error("not allowed", { cause: 401 }));
+    return next(new AppError("not allowed",  401 ));
   }
 
   await comment.deleteOne();
